refactor: implement different types of btn
