+++ lib/auto_updater_overlay.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:my_wit_wallet/constants.dart';
 import 'package:my_wit_wallet/theme/extended_theme.dart';
 import 'package:my_wit_wallet/util/get_localization.dart';
-import 'package:my_wit_wallet/widgets/PaddedButton.dart';
+import 'package:my_wit_wallet/widgets/buttons/custom_btn.dart';
 import 'package:updat/updat.dart';
 import 'package:updat/updat_window_manager.dart';
