+++ lib/widgets/balance.dart
@@ -9,7 +9,7 @@ import 'package:my_wit_wallet/theme/extended_theme.dart';
 import 'package:my_wit_wallet/util/get_localization.dart';
 import 'package:my_wit_wallet/util/storage/database/account.dart';
 import 'package:my_wit_wallet/util/storage/database/wallet.dart';
-import 'package:my_wit_wallet/widgets/PaddedButton.dart';
+import 'package:my_wit_wallet/widgets/buttons/icon_btn.dart';
 import 'package:my_wit_wallet/widgets/copy_button.dart';
 import 'package:my_wit_wallet/util/extensions/string_extensions.dart';
 import 'package:my_wit_wallet/util/extensions/num_extensions.dart';
@@ -63,11 +63,11 @@ class BalanceState extends State<Balance> {
                                     style: theme.textTheme.titleLarge!.copyWith(
                                         color: extendedTheme.headerTextColor)),
                                 Flexible(
-                                  child: PaddedButton(
+                                  child: IconBtn(
                                       padding: EdgeInsets.zero,
                                       label: localization.showBalanceDetails,
                                       text: localization.showBalanceDetails,
-                                      type: ButtonType.iconButton,
+                                      iconBtnType: IconBtnType.icon,
                                       iconSize: 12,
                                       onPressed: widget.onShowBalanceDetails,
                                       icon: Icon(
