+++ lib/widgets/generate_compatible_xprv.dart
@@ -4,7 +4,8 @@ import 'package:my_wit_wallet/util/get_localization.dart';
 import 'package:my_wit_wallet/bloc/crypto/api_crypto.dart';
 import 'package:my_wit_wallet/shared/api_database.dart';
 import 'package:my_wit_wallet/shared/locator.dart';
-import 'package:my_wit_wallet/widgets/PaddedButton.dart';
+
+import 'package:my_wit_wallet/widgets/buttons/custom_btn.dart';
 import 'package:my_wit_wallet/widgets/input_login.dart';
 import 'package:flutter/material.dart';
 import 'package:my_wit_wallet/widgets/labeled_form_entry.dart';
@@ -183,10 +184,10 @@ class GenerateCompatibleXprvState extends State<GenerateCompatibleXprv>
                     },
                   )),
               SizedBox(height: 16),
-              PaddedButton(
+              CustomButton(
                   padding: EdgeInsets.only(bottom: 8),
                   text: localization.generateXprv,
-                  type: ButtonType.primary,
+                  type: CustomBtnType.primary,
                   isLoading: isLoading,
                   enabled: true,
                   onPressed: () async {
